chanstate: move commit diff serialization
Move commit diff serialization and the commitment auxiliary TLV helpers
into chanstate so pending remote commitment state no longer depends on
channeldb codecs. Update channeldb callers to use the chanstate helpers
while leaving open-channel bucket persistence in place for later
backend-specific splits.
